+++ CHANGELOG.md
@@ -2,6 +2,51 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+# [4.3.0](https://github.com/quintype/quintype-node-framework/compare/v4.2.8...v4.3.0) (2020-10-13)
+
+
+### Features
+
+* **amp:** AMP Subscriptions ([768d65d](https://github.com/quintype/quintype-node-framework/commit/768d65d))
+
+
+
+## [4.2.8](https://github.com/quintype/quintype-node-framework/compare/v4.2.7...v4.2.8) (2020-10-12)
+
+
+### Bug Fixes
+
+* **redirectUrls:** Pass config to redirect urls function ([#177](https://github.com/quintype/quintype-node-framework/issues/177)) ([8620962](https://github.com/quintype/quintype-node-framework/commit/8620962))
+
+
+
+## [4.2.7](https://github.com/quintype/quintype-node-framework/compare/v4.2.6...v4.2.7) (2020-10-12)
+
+
+### Bug Fixes
+
+* **amp:** pass only domain specific opts to amp lib ([#175](https://github.com/quintype/quintype-node-framework/issues/175)) ([2d01fd6](https://github.com/quintype/quintype-node-framework/commit/2d01fd6))
+
+
+
+## [4.2.6](https://github.com/quintype/quintype-node-framework/compare/v4.2.5...v4.2.6) (2020-10-06)
+
+
+### Bug Fixes
+
+* **amp:** bump amplib to take header backgroundColor fix ([#173](https://github.com/quintype/quintype-node-framework/issues/173)) ([8bebc54](https://github.com/quintype/quintype-node-framework/commit/8bebc54))
+
+
+
+## [4.2.5](https://github.com/quintype/quintype-node-framework/compare/v4.2.4...v4.2.5) (2020-10-06)
+
+
+### Bug Fixes
+
+* **amp:** update amplib to get heroimage enlarge SEO fix ([#172](https://github.com/quintype/quintype-node-framework/issues/172)) ([ccd666c](https://github.com/quintype/quintype-node-framework/commit/ccd666c))
+
+
+
 ## [4.2.4](https://github.com/quintype/quintype-node-framework/compare/v4.2.3...v4.2.4) (2020-09-23)